% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_table}
\alias{feature_table}
\title{feature_table}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 23 rows and 73 columns.
}
\usage{
feature_table
}
\description{
feature_table
}
\keyword{datasets}
