% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{stamp R6 class}
\description{
stamp R6 class
}
\details{
stamp R6 class
}
\examples{

## ------------------------------------------------
## Method `stamp$new`
## ------------------------------------------------

obj <- stamp$new(feature_table = feature_table, metadata = metadata)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature_table}}{一个data.frame或者tibble，用来存储特征矩阵
第一列是特征列，如基因或者微生物等}

\item{\code{metadata}}{一个data.frame或者tibble，用来存储分组信息
第一列是样本列}

\item{\code{diff}}{一个data.frame或者tibble，用来存储差异分析结果}

\item{\code{diff_features}}{一个字符向量，用来存储差异特征}

\item{\code{test}}{差异分析使用的假设检验，默认是wilcox.test}

\item{\code{feature_col}}{特征列名}

\item{\code{sample_col}}{样本列名}

\item{\code{group_col}}{分组列名}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{longdata}}{long data}

\item{\code{stats}}{stats}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stamp-new}{\code{stamp$new()}}
\item \href{#method-stamp-cal_diff}{\code{stamp$cal_diff()}}
\item \href{#method-stamp-filter_diff_features}{\code{stamp$filter_diff_features()}}
\item \href{#method-stamp-plot_extended_errorbar}{\code{stamp$plot_extended_errorbar()}}
\item \href{#method-stamp-clone}{\code{stamp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stamp-new"></a>}}
\if{latex}{\out{\hypertarget{method-stamp-new}{}}}
\subsection{Method \code{new()}}{
Create a new stamp object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stamp$new(feature_table = NA, metadata = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_table}}{feature_table.}

\item{\code{metadata}}{metadata.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{stamp} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj <- stamp$new(feature_table = feature_table, metadata = metadata)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stamp-cal_diff"></a>}}
\if{latex}{\out{\hypertarget{method-stamp-cal_diff}{}}}
\subsection{Method \code{cal_diff()}}{
cal diff
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stamp$cal_diff(test = "wilcox.test", p.adjust.method = "fdr")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test}}{default="wilcox.test" "t.test"}

\item{\code{p.adjust.method}}{default="fdr"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data.frame or tibble, diff test result
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stamp-filter_diff_features"></a>}}
\if{latex}{\out{\hypertarget{method-stamp-filter_diff_features}{}}}
\subsection{Method \code{filter_diff_features()}}{
filter diff features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stamp$filter_diff_features(p_threshold = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p_threshold}}{default=0.05}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
null
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stamp-plot_extended_errorbar"></a>}}
\if{latex}{\out{\hypertarget{method-stamp-plot_extended_errorbar}{}}}
\subsection{Method \code{plot_extended_errorbar()}}{
plot extended errorbar plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stamp$plot_extended_errorbar()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stamp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-stamp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stamp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
