% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp4r-package.R
\docType{package}
\name{stamp4r-package}
\alias{stamp4r}
\alias{stamp4r-package}
\title{stamp4r: stamp4r是一个分析不同组间特征差异并绘图展示的软件包}
\description{
stamp4r是一个分析不同组间特征差异并绘图展示的软件包.
}
\author{
\strong{Maintainer}: wang zhichao \email{wangzhichao1990@126.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
